requirements.txt
Flask==2.2.5
mysql-connector-python==8.0.34
pandas==2.1.1
plotly==5.16.1
scikit-learn==1.3.1
prophet==1.1